+++ AssetsFlow.Client.React/src/Components/Journal/TradeComposite/Controls/TradeClosingModal.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Modal, NumberInput, Button } from '@mantine/core';
 import { useForm } from '@mantine/form';