+++ AssetsFlow.Client.React/src/Components/Journal/TradeElement/TradeElementWrapper.jsx
@@ -55,7 +55,6 @@ function TradeElementWrapper({ tradeElement, onElementContentUpdate, onElementAc
         </ActionIcon>
       </div>
 
-      {/* Render content */}
       {isCollapsed ? (
         <TradeElementCollapsed tradeElement={tradeElement}/>
       ) : (