+++ AssetsFlow.Server/AssetsFlow.Web.Services/Mapping/EntitiesToDtosMappingProfile.cs
@@ -15,7 +15,13 @@ namespace HsR.Web.API.Mapping
 
             CreateMap<DataElement, DataElementModel>().EqualityComparison((dto, m) => dto.Id == m.Id);
             CreateMap<TradeElement, TradeElementModel>().EqualityComparison((dto, m) => dto.Id == m.Id);
-            CreateMap<TradeComposite, TradeCompositeModel>().EqualityComparison((dto, m) => dto.Id == m.Id);
+
+            CreateMap<TradeComposite, TradeCompositeModel>()
+                .EqualityComparison((dto, m) => dto.Id == m.Id)
+                .ForMember(dest => dest.IsAnyContentMissing, 
+                opt => opt.MapFrom(src => src.TradeElements.Any(ele => !ele.AllowActivation())
+                ));
+
 
             CreateMap<UpdatedStatesCollation, UpdatedStatesModel>();
             CreateMap<TradeComposite, TradeCompositeInfo>().EqualityComparison((dto, m) => dto.Id == m.Id);