+++ AssetsFlow.Server/AssetsFlow.Web.Services/Models/Journal/TradeCompositeModel.cs
@@ -16,7 +16,8 @@ namespace HsR.Web.Services.Models.Journal
         [JsonConverter(typeof(JsonStringEnumConverter))]
         public TradeStatus Status { get; set; }
 
-        public bool IsPending { get; set; }
+        [Required]
+        public bool IsAnyContentMissing { get; set; }
 
         public DateTime? OpenedAt { get; set; }
         public DateTime? ClosedAt { get; set; }