+++ AssetsFlow.Client.React/CHANGELOG.md
@@ -1,6 +1,13 @@
 
 # Changelog
 
+## [0.9.013-f] - 2025-06-20
+### Changed
+- IsPending identifier string due to change in the BE
+### Added
+- Trade notification control - show on missing content
+- Trade admin controls - under dev for now. 
+
 ## [0.9.12] - 2025-05-26
 ### Added
 - Red border on missing data, when you open a trade